requirements.txt
# Telegram libraries
telethon==1.34.0
python-telegram-bot==21.6

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Scheduling
apscheduler==3.10.4

# Environment & Configuration
python-dotenv==1.0.0

# Async utilities
aiohttp==3.9.1

# Timezone handling
pytz==2023.3

# Local AI (Ollama)
ollama==0.3.3

